the client side code is fully functional in the sense that it will upload a pho to to content server, display a spinner while it is doing it, show a success mes sage when it comes back, and then hide the spinner and success message after a t imeout of three seconds.
right now, the username and password are hardcoded in. im working to use either the otds ticket or some other authentication mechanism. the trouble is getting a hold of the otds ticket since its not included with the otag stuff.

also working to move this content server instance to the one we will be using for the demo.